relocate kbd & mouse
